requirements.txt
fastapi
uvicorn
fastapi[standard]
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client